pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ctf-challenge-builder"
version = "0.2.23"
description = "Build challenge images and create OCI deployment packages for CTF challenges"
readme = "README.md"
license = "MIT"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "PyYAML>=6.0",
    "colorama>=0.4.6",
    "requests>=2.25.0",
]

[project.scripts]
build-challenge = "ctf_challenge_builder:main"

[tool.setuptools]
packages = ["ctf_challenge_builder"]

[tool.setuptools.package-data]
ctf_challenge_builder = ["pulumi-template/*"]